Use Primary as readPreference
We observed some records coming through with empty data other than _id.
It seems plausible that this is because it is being read off of oplog
but not yet present in replicas. If we read off a replica, we would not
yet see the information.

This commit changes the setting from Monotonic
```
https://github.com/go-mgo/mgo/blob/v2/session.go#L62
Same as SecondaryPreferred before first write. Same as Primary after first write.
```

To Primary.