Update GTM and rework system to deal with changes
Upgrading to GTM: SHA e400ebf9f055864f277fcb1471f851a4c112b603
fixes issue where multiple results for same key in short period results
in empty rows in Postgres. The rows are inserted with only an _id.

This was because GTM iterated over the results and expected to lack
duplicates in the `this.Entries`. Appears fixed in recent commit.

Moresql required changes to accommodate GTM changes. Namely, it's
necessary to copy the incoming op.Data in order to avoid issues with
concurrent access read/write to maps.