Crash application on Mongo read errors
Mongo read/connectivity issues are causing this application to stop
working silently. Instead of silently receiving them and logging, we
will crash the application and rely on external process monitor to
revive the service.

In future we can consider more graceful recovery mechanism with exiting
`Read()` loop and restarting it when connectivity issues arise.