Add tailing recovery in cases of timeout
In cases of i/o timeout from GTM, moresql needs to re-establish the tail
mechanism.

We do this by establishing the last known epoch timestamp and creating a
new gtm.Tail. Those output channels are attached and will be used in the
next iteration through the loop. In case it's unrecoverable or unknown
error, we fail out and crash application.